pyproject.toml
[project]
name = "e-commerce-minimalist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.1.6",
    "django-decouple>=2.1",
    "django-phonenumber-field>=8.0.0",
    "djangorestframework>=3.15.2",
    "model-bakery>=1.20.1",
    "phonenumbers>=8.13.54",
    "pillow>=11.1.0",
    "psycopg2>=2.9.10",
    "pytest-django>=4.10.0",
]
